requirements.txt
aio-pika==9.4.2
aioredis==1.3.1
aiormq==6.8.0
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==3.2.0
boto3==1.34.158
botocore==1.34.158
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.1
fastapi-cache==0.1.0
fastapi-cli==0.0.5
gcloud==0.18.3
googleapis-common-protos==1.63.2
greenlet==3.0.3
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jwcrypto==1.5.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
oauth2client==4.1.3
orjson==3.10.6
packaging==24.1
pamqp==3.3.0
passlib==1.7.4
pika==1.3.2
pluggy==1.5.0
protobuf==5.28.0rc2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pytest==8.3.2
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-jwt==4.1.0
python-multipart==0.0.9
PyYAML==6.0.1
rabbitmq==0.2.0
redis==5.0.8
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlmodel==0.0.21
starlette==0.38.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.30.5
uvloop==0.20.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
